feat(places): Security

+++ src/main/java/fr/foodlimit/api/place/PlaceController.java
@@ -4,6 +4,8 @@ import fr.foodlimit.api.security.jwt.JWTFilter;
 import fr.foodlimit.api.security.jwt.TokenProvider;
 import fr.foodlimit.api.shared.models.Place;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
@@ -34,8 +36,8 @@ public class PlaceController {
    * @return
    */
   @GetMapping
-  public List<Place> getPlaces(HttpServletRequest request) {
-    return placeService.getPlaces(tokenProvider.getUsername(JWTFilter.resolveToken(request)));
+  public ResponseEntity<List<Place>> getPlaces(HttpServletRequest request) {
+    return ResponseEntity.ok(placeService.getPlaces(tokenProvider.getUsername(JWTFilter.resolveToken(request))));
   }
 
   /**
@@ -44,8 +46,15 @@ public class PlaceController {
    * @return
    */
   @GetMapping("/{id}")
-  public Place getPlace(@PathVariable("id") Long id) {
-    return placeService.getPlace(id);
+  public ResponseEntity<Place> getPlace(HttpServletRequest request, @PathVariable("id") Long id) {
+    Place place = placeService.getPlace(id);
+    String username = tokenProvider.getUsername(JWTFilter.resolveToken(request));
+
+    if (!place.getUser().getUsername().equals(username)) {
+      return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+    }
+
+    return ResponseEntity.ok(placeService.getPlace(id));
   }
 
   /**
@@ -53,8 +62,16 @@ public class PlaceController {
    * @param id
    */
   @DeleteMapping("/{id}")
-  public void deletePlace(@PathVariable("id") Long id) {
+  public ResponseEntity deletePlace(HttpServletRequest request, @PathVariable("id") Long id) {
+    Place place = placeService.getPlace(id);
+    String username = tokenProvider.getUsername(JWTFilter.resolveToken(request));
+
+    if (!place.getUser().getUsername().equals(username)) {
+      return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+    }
+
     placeService.deletePlace(id);
+    return ResponseEntity.noContent().build();
   }
 
   /**
@@ -64,8 +81,8 @@ public class PlaceController {
    * @return
    */
   @PostMapping
-  public Place createPlace(HttpServletRequest request, @RequestBody Place place) {
-    return placeService.createPlace(place, tokenProvider.getUsername(JWTFilter.resolveToken(request)));
+  public ResponseEntity<Place> createPlace(HttpServletRequest request, @RequestBody Place place) {
+    return ResponseEntity.ok(placeService.createPlace(place, tokenProvider.getUsername(JWTFilter.resolveToken(request))));
   }
 
   /**
@@ -76,9 +93,16 @@ public class PlaceController {
    * @return
    */
   @PutMapping("/{id}")
-  public Place updatePlace(HttpServletRequest request, @PathVariable("id") Long id, @RequestBody Place place) {
+  public ResponseEntity<Place> updatePlace(HttpServletRequest request, @PathVariable("id") Long id, @RequestBody Place place) {
+    Place dbPlace = placeService.getPlace(id);
+    String username = tokenProvider.getUsername(JWTFilter.resolveToken(request));
+
+    if (!dbPlace.getUser().getUsername().equals(username)) {
+      return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+    }
+
     place.setId(id);
-    return placeService.updatePlace(place,tokenProvider.getUsername(JWTFilter.resolveToken(request)));
+    return ResponseEntity.ok(placeService.updatePlace(place,tokenProvider.getUsername(JWTFilter.resolveToken(request))));
   }
 }
